Move interlacing logic to separate helper
